fix(vite-plugin-nitro): use virtual files for SSR/client renderers
