--- packages/vite-plugin-nitro/src/lib/runtime/renderer-client.ts
@@ -1,8 +1,0 @@
-import { eventHandler } from 'h3';
-
-// @ts-ignore
-import template from '#analog/index';
-
-export default eventHandler(async () => {
-  return template;
-});
