--- packages/vite-plugin-nitro/src/lib/runtime/renderer.ts
@@ -1,20 +1,0 @@
-import { eventHandler, getResponseHeader } from 'h3';
-// @ts-ignore
-import renderer from '#analog/ssr';
-// @ts-ignore
-import template from '#analog/index';
-
-export default eventHandler(async (event) => {
-  const noSSR = getResponseHeader(event, 'x-analog-no-ssr');
-
-  if (noSSR === 'true') {
-    return template;
-  }
-
-  const html = await renderer(event.node.req.url, template, {
-    req: event.node.req,
-    res: event.node.res,
-  });
-
-  return html;
-});
